requirements.txt
Flask==2.1.0
Flask_SQLAlchemy==2.5.1
gunicorn==20.1.0
certifi==2021.10.8
pandas==1.3.5
psycopg2==2.9.3
wincertstore==0.2
numpy==1.21.6